// src/store/types.ts
export enum FiltersActions {
    changeBrightness = '[FILTERS] Change brightness',
    changeContrast = '[FILTERS] Change contrast',
    changeGrayscale = '[FILTERS] Change grayscale',
    changeHueRotate = '[FILTER] Change changeHueRotate',
    changeInvert = '[FILTER] Change changeInvert',
    changeOpacity = '[FILTER] Change changeOpacity',
    changeSaturate = '[FILTER] Change changeSaturate',
    changeSepia = '[FILTER] Change changeSepia',
}